refactor(blockifier): delete transaction_utils
