+++ crates/blockifier/src/transaction.rs
@@ -8,5 +8,4 @@ pub mod objects;
 pub mod test_utils;
 pub mod transaction_execution;
 pub mod transaction_types;
-pub mod transaction_utils;
 pub mod transactions;
