+++ crates/blockifier/src/transaction/transactions.rs
@@ -25,7 +25,7 @@ use crate::execution::entry_point::{
     ConstructorContext,
     EntryPointExecutionContext,
 };
-use crate::execution::execution_utils::execute_deployment;
+use crate::execution::execution_utils::{execute_deployment, update_remaining_gas};
 use crate::state::cached_state::TransactionalState;
 use crate::state::errors::StateError;
 use crate::state::state_api::{State, UpdatableState};
@@ -41,8 +41,6 @@ use crate::transaction::objects::{
     TransactionInfo,
     TransactionInfoCreator,
 };
-use crate::transaction::transaction_utils::{update_remaining_gas, verify_contract_class_version};
-
 #[cfg(test)]
 #[path = "transactions_test.rs"]
 mod test;
@@ -153,7 +151,26 @@ impl DeclareTransaction {
         only_query: bool,
     ) -> TransactionExecutionResult<Self> {
         let declare_version = declare_tx.version();
-        verify_contract_class_version(&class_info.contract_class(), declare_version)?;
+        // Verify contract class version.
+        match &class_info.contract_class() {
+            // TODO: Make TransactionVersion an enum and use match here.
+            ContractClass::V0(_) => {
+                if declare_version > TransactionVersion::ONE {
+                    Err(TransactionExecutionError::ContractClassVersionMismatch {
+                        declare_version,
+                        cairo_version: 0,
+                    })?
+                }
+            }
+            ContractClass::V1(_) => {
+                if declare_version <= TransactionVersion::ONE {
+                    Err(TransactionExecutionError::ContractClassVersionMismatch {
+                        declare_version,
+                        cairo_version: 1,
+                    })?
+                }
+            }
+        }
         Ok(Self { tx: declare_tx, tx_hash, class_info, only_query })
     }
